Fix message resolution
Previous implementation presumed that the edges would be ordered, use a lookup map instead.
